bandaid fix for TeX issues
Someone found that their tex install doesn't like to play nicely with matplotlib. For now I've put a band-aid on the problem: if the computer can't find tex, they use a default matplotlib font. This results in OK but not great pictures, but it's all I can do RN. Also, without latex the combo movies where I plot the power spectrum will really look bad, so I need a better fix for this down the line.