fix builtins & add .npy save
All the builtin demo scripts are fixed (but more debugging is needed to see why it defaults to kdv when it doesn't recognize the input model_kw). Also added the ability to save as .npy rather than .[kl: hopefully this can improve storage restrictions. Maybe will come back and rename the syntax to make the different kinds of saving as clear as possible.... should saving as a .pkl be "save" or "save_pkl"???? In a sense this should be the default but maybe it's also best to keep the _pkl around, or maybe it just doesn't matter all that much (esp. when it comes to code clarity and readability)