fixed sponge syntax
Finally got rid of all mentions of "absorbing_layer" or "abslayer" in favour of just clearly saying "sponge_layer" everywhere! This should make the code a lot more readable and easy to maintain. 

To ensure that I can keep all my old data (especially the agonizing ST simulations....), there's a little caveat in sim.load() that looks for the old file naming convention (with the ugly "abslayer" in it) as well as the new one. This must be deprecated in a future version! Wait until the paper that uses all these sims is published!!!!