optimize fft of real fields
Fixed things so that, during all simulations, the fft of any real field is computed efficiently with scipy.fft.rfft (ditto for inverse fft). Complex-valued fields are left the same as they were before. So, no more extraneous frequencies are computed! I haven't noticed much of a performance boost unfortunately but it's something that a nasty person could ask about in a conference/peer review so tbh glad I just buckled down and fixed it. 

Also added a file "utils.py" for taking a wrapped fft/ifft where complexity can be toggled, and where the function "integrate" now lives. I didn't change the fft syntax on "integrate" bcz it's for postprocessing so speed isn't such a big deal with this guy.